Add testing of the linear unit.
After commiting the linear unit, I needed to test it. I created an XPP program, sim4a.ode, that is a modified version
of sim3b.ode, with the only difference of using linear units instead of sigmoidal ones. I ran sim4a.ode and saved the
results in sim4aoderun.dat .

I created a copy of test1.ipynb and modified the cell for TEST 3 so it used linear units, and compared the results with
those of sim4aoderun.dat . The results match :)