Now with physical plants.
I have finished the implementation of the parent 'plant' class for
physical plant models, and of the first particular model, which
is the 'pendulum' class.

I modified network.create() and network.run() so it could create
instances of plants, and would run them along with the units.
I still haven't written code to connect the plants with the units.

In order to test 'pendulum' I wrote test6.ipynb, which compares a
the output of a simulation, with that of the same simulation done in
XPPAUT. The XPP simulation was from pend.ode, which is an example
included in the XPP distribution.