Add function to connect inputs to a plant.
I implemented the function network.set_plant_inputs, and made several
additions to the plant class so this could work. Plants now have the
structures inputs, inp_dels, and inp_syns, which take the role of
the network lists act, delays, and syns, but with the possibility of
multiple input ports.

I modified test6 to test set_plant_inputs. The test using the pendulum
model was successful. I still need to test with a model that has
more than one input port, so I created a 3-dim model called conn_test, with
very simple dynamics, and 3 input ports. I'm in the middle of writing
a cell in test6 that puts inputs in conn_test and runs it.w