Introduced main modifications for simulation backup in mp_net_runner.
I've modified the mp_net_runner class in order to put the option of
saving every save_interv presentations in run_all. Doing it in a manner
I liked required modifying the run_net method so that its argument
became a 2-tuple with the net number and the number of presentations.
This change could break past compatibility in two ways:
* Any old mp_net_runner objects that you load don't have the new run_all
  interface, so you shouldn't invoke it with the save_interv argument.
* Saving affects which patters get presented to the net, because the
  n_pres argument to ei_network.run only goes as far as save_interv.