Design for the implementation of requirements is finalized for lpf_fast.
This commit shows the result of a long series of efforts to find a
good way to improve the implementation of requirements, removing the
wall of elif statements in init_pre_syn_update. Eventually two options
were tested, and the nicest one had to be discarded because it slows
things down. This is described in the notes on init_pre_syn_update.

I might get rid of 'requirements.py' in the future.