I fixed the third flat version of the networ. It is slightly slower than the second, but it makes up in its distributed structure.
In order to prepare for extension of the flattening to various
models, I modified the code for units so that a new 'integ_meth'
parameter is accepted. This parameter selects the integration
mehtod. I have created a separate function for each of the
integration methods, and they are all part of the unit class.

I was implementing a similar thing for the flat network, where
the flatten method specifies a 'flat_update' method for the unit,
but I ran out of time.