Configuration good without static fusimotor drive.
This configuration has the modelling approach in Shadmehr and Wise. It
seems sensible until you apply fusimotor drive to the muscle, which
contracts the polar region. Since the II output is proportional to the
length of the polar region, static fusimotor drive reduces II output
rather than increasing it, which is terrible.