Add new integ_decay_act requirement.
This is currently being used in the gated_rga_inpsel_adapt_sig units
in order to replace the lpf_slow component as the integral part
of the control cycle.

Moreover, the C parameters in the base_v3-bp, base_v3-ta_bp, and
base_v3-log_bp notebooks have changed.