WIP: multilayer implementation of RGA.
I have created several new synpse and unit types in preparation of this.

Currently working on the diff_inp_corr_synapse.
It is a curious problem to find the index of the synapse in the inp_deriv_mp
list. To this end, all synapses have a new 'syns_loc' attribute, which I hope
will help.