Two main improvements to the model.
There are a lot of new analysis cells in reach_analysis_D, to the point
where it is already enough to produce the paper results when elements
are removed from the system.

Another important change is that the STATIC_CONN configuration no longer
has plastic A__C and A__M synapses. This required a new A__CM_rand_w
parameter for net_from_cfg in net_builders.